Move ResultsButtons to shared components. Add configuration for UniParc download.
