+++ src/shared/config/resultsDownload.ts
@@ -1,9 +1,11 @@
 import { FileFormat, ContentType } from '../types/resultsDownload';
 import { fileFormatsResultsDownload as fileFormatsResultsDownloadUniProtKB } from '../../uniprotkb/config/download';
 import { fileFormatsResultsDownload as fileFormatsResultsDownloadUniRef } from '../../uniref/config/download';
+import { fileFormatsResultsDownload as fileFormatsResultsDownloadUniParc } from '../../uniparc/config/download';
 import { Namespace } from '../types/namespaces';
 
 export const fileFormatToContentType: Record<FileFormat, ContentType> = {
+  [FileFormat.fasta]: ContentType.fasta,
   [FileFormat.fastaCanonical]: ContentType.fasta,
   [FileFormat.fastaCanonicalIsoform]: ContentType.fasta,
   [FileFormat.fastaRepresentative]: ContentType.fasta,
@@ -18,6 +20,7 @@ export const fileFormatToContentType: Record<FileFormat, ContentType> = {
 };
 
 export const fileFormatToUrlParameter: Record<FileFormat, string> = {
+  [FileFormat.fasta]: 'fasta',
   [FileFormat.fastaCanonical]: 'fasta',
   [FileFormat.fastaCanonicalIsoform]: 'fasta',
   [FileFormat.fastaRepresentative]: 'fasta',
@@ -39,4 +42,5 @@ export const nsToFileFormatsResultsDownload: Partial<
 > = {
   [Namespace.uniprotkb]: fileFormatsResultsDownloadUniProtKB,
   [Namespace.uniref]: fileFormatsResultsDownloadUniRef,
+  [Namespace.uniparc]: fileFormatsResultsDownloadUniParc,
 };
